backend/requirements.txt
fastapi
uvicorn
appwrite
python-dotenv
pydantic
pydantic-settings
requests
google-generativeai